+++ src/process_handler.py
@@ -2,20 +2,7 @@ import asyncio
 import pwd
 import os
 import signal
-
-import aiofiles
-
-IPC_CHANNELS = [
-    'video'
-]
-
-
-def get_cmd_prefix():
-    # run as pi user if available
-    for user in pwd.getpwall():
-        if user[0] == 'pi':
-            return 'sudo -u pi '
-    return ''
+from shlex import split
 
 
 class InvalidOperation(Exception):
@@ -23,11 +10,10 @@ class InvalidOperation(Exception):
 
 
 class ProcessHandler:
-    def __init__(self, on_start, on_stop, on_output, work_dir='/tmp'):
+    def __init__(self, on_start, on_stop, on_output):
         self.on_start = on_start
         self.on_stop = on_stop
         self.on_output = on_output
-        self.work_dir = work_dir
 
         self.id = str(id(self))
         self.process = None
@@ -36,17 +22,12 @@ class ProcessHandler:
         if self.is_running():
             self.stop()
 
-    async def start(self, script):
-        if self.is_running() or not isinstance(script, str):
+    async def start(self, command):
+        if self.is_running() or not isinstance(command, str):
             raise InvalidOperation()
 
-        main_filename = self._get_main_filename()
-        async with aiofiles.open(main_filename, 'w+') as file:
-            await file.write(script)
-
-        command = get_cmd_prefix() + 'python3 -u ' + main_filename
         self.process = await asyncio.create_subprocess_exec(
-            *command.split(),
+            *split(command),
             stdin=asyncio.subprocess.PIPE,
             stdout=asyncio.subprocess.PIPE,
             stderr=asyncio.subprocess.PIPE,
@@ -73,36 +54,18 @@ class ProcessHandler:
         self.process.stdin.write(content.encode('utf-8'))
         await self.process.stdin.drain()
 
-    def _get_main_filename(self):
-        return self.work_dir + '/' + self.id + '.py'
-
-    def _get_ipc_filename(self, channel):
-        return self.work_dir + '/' + self.id + '.' + channel + '.sock'
-
     async def _communicate(self):
         output_tasks = [
             asyncio.create_task(self._handle_output('stdout')),
             asyncio.create_task(self._handle_output('stderr')),
         ]
 
-        ipc_tasks = []
-        for name in IPC_CHANNELS:
-            ipc_tasks.append(asyncio.create_task(
-                self._handle_ipc(name)
-            ))
-
         # allow the output tasks to finish & flush
         await asyncio.wait(output_tasks)
-
-        # stop ongoing ipc servers
-        for task in ipc_tasks:
-            task.cancel()
-        await asyncio.wait(ipc_tasks)
 
         # process should be done now but await it to get exit code
         exit_code = await self.process.wait()
         self.process = None
-        await self._clean_up()
 
         if self.on_stop:
             await self.on_stop(exit_code)
@@ -117,35 +80,3 @@ class ProcessHandler:
             output = data.decode(encoding='utf-8')
             if self.on_output:
                 await self.on_output(stream_name, output)
-
-    async def _handle_ipc(self, channel):
-        async def handle_connection(reader, _):
-            message = ''
-            while True:
-                data = await reader.read(4096)
-                if data == b'':
-                    break
-
-                tokens = data.decode('utf-8').strip().split()
-                if tokens[0] == channel:
-                    if len(message) > 0:
-                        if self.on_output:
-                            await self.on_output(channel, message)
-                        message = ''
-                    message += tokens[1]
-                else:
-                    message += tokens[0]
-
-        ipc_filename = self._get_ipc_filename(channel)
-        await asyncio.start_unix_server(handle_connection, path=ipc_filename)
-
-    async def _clean_up(self):
-        try:
-            await aiofiles.os.remove(self._get_main_filename())
-            for name in ipc_channel_names:
-                try:
-                    await aiofiles.os.remove(self._get_ipc_filename(name))
-                except:
-                    pass
-        except:
-            pass
